feat(infra): add original doc access method
