--- tests/console/app_container_test.py
@@ -1,20 +1,0 @@
-"""
-test global chk functions
-"""
-from chk.infrastructure.containers import App
-from chk.infrastructure.helper import data_set, data_get
-
-app = App()
-
-
-class TestChk:
-    """Test chk functions"""
-
-    @staticmethod
-    def test_app_variables_set():
-        data_set(app.original_doc, "some.key", "Some name: {name}")
-        assert app.original_doc["some"]["key"] == "Some name: {name}"
-
-    @staticmethod
-    def test_app_variables_get():
-        assert data_get(app.original_doc, "some.key") == "Some name: {name}"
